requirements.txt
requests
django
pillow
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt
django-filters